requirements.txt
torch
neuraloperator
numpy<2
matplotlib
h5py
